refactor: 정적 데이터 엔티티 폴더 위치 변경
