+++ src/main/java/com/odiga/fiesta/global/domain/Priority.java
@@ -1,4 +1,4 @@
-package com.odiga.fiesta.festival.domain;
+package com.odiga.fiesta.global.domain;
 
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
