+++ src/main/java/com/odiga/fiesta/global/repository/CompanionRepository.java
@@ -1,8 +1,8 @@
-package com.odiga.fiesta.festival.repository;
+package com.odiga.fiesta.global.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.odiga.fiesta.festival.domain.Companion;
+import com.odiga.fiesta.global.domain.Companion;
 
 public interface CompanionRepository extends JpaRepository<Companion, Long> {
 }
