+++ src/main/java/com/odiga/fiesta/global/repository/MoodRepository.java
@@ -1,8 +1,8 @@
-package com.odiga.fiesta.festival.repository;
+package com.odiga.fiesta.global.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.odiga.fiesta.festival.domain.Mood;
+import com.odiga.fiesta.global.domain.Mood;
 
 public interface MoodRepository extends JpaRepository<Mood, Long> {
 }
