+++ src/main/java/com/odiga/fiesta/global/service/PriorityService.java
@@ -1,13 +1,13 @@
-package com.odiga.fiesta.festival.service;
+package com.odiga.fiesta.global.service;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.odiga.fiesta.festival.domain.Priority;
+import com.odiga.fiesta.global.domain.Priority;
 import com.odiga.fiesta.festival.dto.response.PriorityResponse;
-import com.odiga.fiesta.festival.repository.PriorityRepository;
+import com.odiga.fiesta.global.repository.PriorityRepository;
 
 import lombok.RequiredArgsConstructor;
