+++ src/test/java/com/odiga/fiesta/festival/controller/FestivalControllerTest.java
@@ -16,10 +16,10 @@ import com.odiga.fiesta.festival.dto.response.CategoryResponse;
 import com.odiga.fiesta.festival.dto.response.CompanionResponse;
 import com.odiga.fiesta.festival.dto.response.MoodResponse;
 import com.odiga.fiesta.festival.dto.response.PriorityResponse;
-import com.odiga.fiesta.festival.service.CategoryService;
-import com.odiga.fiesta.festival.service.CompanionService;
-import com.odiga.fiesta.festival.service.MoodService;
-import com.odiga.fiesta.festival.service.PriorityService;
+import com.odiga.fiesta.global.service.CategoryService;
+import com.odiga.fiesta.global.service.CompanionService;
+import com.odiga.fiesta.global.service.MoodService;
+import com.odiga.fiesta.global.service.PriorityService;
 
 class FestivalControllerTest extends ControllerTestSupport {
