+++ src/test/java/com/odiga/fiesta/festival/service/PriorityServiceTest.java
@@ -11,9 +11,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
 import com.odiga.fiesta.MockTestSupport;
-import com.odiga.fiesta.festival.domain.Priority;
+import com.odiga.fiesta.global.domain.Priority;
 import com.odiga.fiesta.festival.dto.response.PriorityResponse;
-import com.odiga.fiesta.festival.repository.PriorityRepository;
+import com.odiga.fiesta.global.repository.PriorityRepository;
+import com.odiga.fiesta.global.service.PriorityService;
 
 class PriorityServiceTest extends MockTestSupport {
